# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bhulan GPS API",
    "description": "Mobility insights and map plotting for GPS coordinates, plus ingestion endpoints for vendor feeds.",
    "version": "2.1.0"
  },
  "paths": {
    "/v1/insights": {
      "post": {
        "tags": [
          "insights"
        ],
        "summary": "Insights Endpoint",
        "description": "Compute mobility insights for a batch of GPS coordinates.\n\nAccepts either a structured ``points`` array (fast path) or raw ``text``\nthat the server will parse using :func:`bhulan.analytics.parsers.parse_any`.\nSetting ``options.geocode_stops = true`` enriches each detected stop\nwith a Nominatim-derived ``place_name`` (adds ~1s per unique stop).\n\nWhen the caller is authenticated (``Authorization: Bearer <session>``)\nand server-side auth is enabled, the run is also persisted to the\nuser's history \u2014 look it up with ``GET /v1/history``.",
        "operationId": "insights_endpoint_v1_insights_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RawInsightsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightsReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/parse/file": {
      "post": {
        "tags": [
          "insights"
        ],
        "summary": "Parse File Endpoint",
        "description": "Parse an uploaded GPS file and return normalized points.\n\nDispatches on the filename extension: ``.gpx`` \u2192 gpxpy, ``.kml`` \u2192\nstdlib XML, ``.fit`` \u2192 fitdecode, anything else \u2192 the text parsers.\nFiles over ``MAX_UPLOAD_BYTES`` are rejected; the content is read into\nmemory once and parsed synchronously (parsing is CPU-bound and fast\nenough that pushing it to a threadpool isn't worth it at this scale).",
        "operationId": "parse_file_endpoint_v1_parse_file_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_parse_file_endpoint_v1_parse_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseFileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/plot/validate": {
      "post": {
        "tags": [
          "insights"
        ],
        "summary": "Plot Validate Endpoint",
        "description": "Parse and validate coordinates for the map view.\n\nReturns the cleaned points plus counts of accepted and rejected rows so\nthe frontend can show a short summary alongside the rendered track.",
        "operationId": "plot_validate_endpoint_v1_plot_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_parse_file_endpoint_v1_parse_file_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "GPX, KML, FIT, CSV, JSON, or plain-text coordinates"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_parse_file_endpoint_v1_parse_file_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HotspotOut": {
        "properties": {
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "sample_count": {
            "type": "integer",
            "title": "Sample Count"
          },
          "visit_count": {
            "type": "integer",
            "title": "Visit Count"
          },
          "time_spent_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Spent Min"
          },
          "first_ts": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Ts"
          },
          "last_ts": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Ts"
          },
          "place_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place Name",
            "description": "Reverse-geocoded label for the centroid. Only populated when InsightsOptions.geocode_stops is true."
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lon",
          "sample_count",
          "visit_count",
          "time_spent_min",
          "first_ts",
          "last_ts"
        ],
        "title": "HotspotOut",
        "description": "One detected density cluster.\n\n``time_spent_min`` is ``None`` when the input has no timestamps \u2014\nwe can still identify the hotspot geographically, we just can't\nsay how long was spent there."
      },
      "InsightsOptions": {
        "properties": {
          "stop_radius_m": {
            "type": "number",
            "maximum": 10000.0,
            "exclusiveMinimum": 0.0,
            "title": "Stop Radius M",
            "default": 50.0
          },
          "min_stop_minutes": {
            "type": "number",
            "maximum": 1440.0,
            "exclusiveMinimum": 0.0,
            "title": "Min Stop Minutes",
            "description": "Minutes",
            "default": 5.0
          },
          "moving_speed_kmh": {
            "type": "number",
            "maximum": 300.0,
            "minimum": 0.0,
            "title": "Moving Speed Kmh",
            "description": "Speed threshold for 'moving'",
            "default": 3.6
          },
          "merge_stops_within_m": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Merge Stops Within M",
            "description": "If set, merge consecutive stops whose centroids are within this radius"
          },
          "geocode_stops": {
            "type": "boolean",
            "title": "Geocode Stops",
            "description": "If true, reverse-geocode each stop's centroid via Nominatim and populate StopOut.place_name. Respects Nominatim's 1 req/sec limit; expect +N seconds latency for N unique stops.",
            "default": false
          },
          "trip_split_stop_minutes": {
            "type": "number",
            "maximum": 1440.0,
            "exclusiveMinimum": 0.0,
            "title": "Trip Split Stop Minutes",
            "description": "Stops longer than this end the current trip and start a new one. Shorter stops stay inside the trip they happen during.",
            "default": 30.0
          },
          "trip_split_gap_minutes": {
            "type": "number",
            "maximum": 10080.0,
            "exclusiveMinimum": 0.0,
            "title": "Trip Split Gap Minutes",
            "description": "A gap between consecutive samples longer than this ends the current trip (device-off or missing data).",
            "default": 60.0
          },
          "hotspot_grid_m": {
            "type": "number",
            "maximum": 10000.0,
            "exclusiveMinimum": 0.0,
            "title": "Hotspot Grid M",
            "description": "Grid cell size (meters) used for hotspot clustering.",
            "default": 100.0
          },
          "hotspot_min_samples": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Hotspot Min Samples",
            "description": "Minimum samples per grid cell to qualify as a hotspot.",
            "default": 5
          },
          "hotspot_max_results": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Hotspot Max Results",
            "description": "Maximum number of hotspots returned, sorted by sample count desc.",
            "default": 10
          }
        },
        "type": "object",
        "title": "InsightsOptions",
        "description": "Tunables for the insights pipeline."
      },
      "InsightsQuality": {
        "properties": {
          "rejected_points": {
            "type": "integer",
            "title": "Rejected Points",
            "default": 0
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "title": "InsightsQuality"
      },
      "InsightsReport": {
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/InsightsSummary"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/StopOut"
            },
            "type": "array",
            "title": "Stops"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/SegmentOut"
            },
            "type": "array",
            "title": "Segments"
          },
          "trips": {
            "items": {
              "$ref": "#/components/schemas/TripOut"
            },
            "type": "array",
            "title": "Trips"
          },
          "hotspots": {
            "items": {
              "$ref": "#/components/schemas/HotspotOut"
            },
            "type": "array",
            "title": "Hotspots"
          },
          "quality": {
            "$ref": "#/components/schemas/InsightsQuality"
          }
        },
        "type": "object",
        "required": [
          "summary",
          "stops",
          "segments",
          "quality"
        ],
        "title": "InsightsReport"
      },
      "InsightsSummary": {
        "properties": {
          "point_count": {
            "type": "integer",
            "title": "Point Count"
          },
          "accepted_point_count": {
            "type": "integer",
            "title": "Accepted Point Count"
          },
          "time_range": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TimeRange"
              },
              {
                "type": "null"
              }
            ]
          },
          "total_distance_km": {
            "type": "number",
            "title": "Total Distance Km"
          },
          "moving_time_min": {
            "type": "number",
            "title": "Moving Time Min"
          },
          "idle_time_min": {
            "type": "number",
            "title": "Idle Time Min"
          },
          "avg_moving_speed_kmh": {
            "type": "number",
            "title": "Avg Moving Speed Kmh"
          },
          "max_speed_kmh": {
            "type": "number",
            "title": "Max Speed Kmh"
          },
          "bbox": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BBox"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "point_count",
          "accepted_point_count",
          "time_range",
          "total_distance_km",
          "moving_time_min",
          "idle_time_min",
          "avg_moving_speed_kmh",
          "max_speed_kmh",
          "bbox"
        ],
        "title": "InsightsSummary"
      },
...
      "ParseFileResponse": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "accepted": {
            "type": "integer",
            "title": "Accepted"
          },
          "rejected": {
            "type": "integer",
            "title": "Rejected"
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/PointIn"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "accepted",
          "rejected",
          "issues",
          "points"
        ],
        "title": "ParseFileResponse",
        "description": "Response for :func:`parse_file_endpoint` \u2014 the frontend feeds the\naccepted points straight into ``/v1/insights`` or ``/v1/plot/validate``."
      },
      "PlotRequest": {
        "properties": {
          "points": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PointIn"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points",
            "description": "Pre-structured points (preferred when the client can parse)"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text",
            "description": "Raw user-pasted text: CSV, JSON, GeoJSON, or lat/lon lines"
          }
        },
        "type": "object",
        "title": "PlotRequest",
        "description": "Request for :func:`plot_validate` \u2014 either structured points or raw text."
      },
      "PlotResponse": {
        "properties": {
          "accepted": {
            "type": "integer",
            "title": "Accepted"
          },
          "rejected": {
            "type": "integer",
            "title": "Rejected"
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/PointIn"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "accepted",
          "rejected",
          "issues",
          "points"
        ],
        "title": "PlotResponse",
        "description": "Normalized points the frontend can drop straight onto a map."
      },
      "PointIn": {
        "properties": {
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon"
          },
          "ts_utc": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ts Utc",
            "description": "Timestamp in UTC; required for any time-based insight"
          },
          "speed_mps": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Mps",
            "description": "Optional device-reported speed"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lon"
        ],
        "title": "PointIn",
        "description": "One GPS sample as it arrives over the wire."
      },
      "RawInsightsRequest": {
        "properties": {
          "points": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PointIn"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "options": {
            "$ref": "#/components/schemas/InsightsOptions"
          },
          "label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Label",
            "description": "Optional label stored alongside the history entry"
          }
        },
        "type": "object",
        "title": "RawInsightsRequest",
        "description": "``/v1/insights`` request that accepts pasted text in lieu of structured input."
      },
      "SegmentOut": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "start_ts": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Ts"
          },
          "end_ts": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Ts"
          },
          "distance_km": {
            "type": "number",
            "title": "Distance Km"
          },
          "duration_min": {
            "type": "number",
            "title": "Duration Min"
          },
          "avg_speed_kmh": {
            "type": "number",
            "title": "Avg Speed Kmh"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "start_ts",
          "end_ts",
          "distance_km",
          "duration_min",
          "avg_speed_kmh"
        ],
        "title": "SegmentOut"
      },
      "StopOut": {
        "properties": {
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "start_ts": {
            "type": "string",
            "format": "date-time",
            "title": "Start Ts"
          },
          "end_ts": {
            "type": "string",
            "format": "date-time",
            "title": "End Ts"
          },
          "duration_min": {
            "type": "number",
            "title": "Duration Min"
          },
          "radius_m": {
            "type": "number",
            "title": "Radius M"
          },
          "sample_count": {
            "type": "integer",
            "title": "Sample Count"
          },
          "place_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place Name",
            "description": "Reverse-geocoded label for the centroid. Only populated when InsightsOptions.geocode_stops is true; otherwise None."
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lon",
          "start_ts",
          "end_ts",
          "duration_min",
          "radius_m",
          "sample_count"
        ],
        "title": "StopOut"
      },
      "TimeRange": {
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          }
        },
        "type": "object",
        "required": [
          "start",
          "end"
        ],
        "title": "TimeRange"
      },
      "TripOut": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index",
            "description": "0-based index of the trip in start order"
          },
          "start_ts": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Ts"
          },
          "end_ts": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Ts"
          },
          "start_lat": {
            "type": "number",
            "title": "Start Lat"
          },
          "start_lon": {
            "type": "number",
            "title": "Start Lon"
          },
          "end_lat": {
            "type": "number",
            "title": "End Lat"
          },
          "end_lon": {
            "type": "number",
            "title": "End Lon"
          },
          "distance_km": {
            "type": "number",
            "title": "Distance Km"
          },
          "duration_min": {
            "type": "number",
            "title": "Duration Min"
          },
          "moving_time_min": {
            "type": "number",
            "title": "Moving Time Min"
          },
          "idle_time_min": {
            "type": "number",
            "title": "Idle Time Min"
          },
          "max_speed_kmh": {
            "type": "number",
            "title": "Max Speed Kmh"
          },
          "sample_count": {
            "type": "integer",
            "title": "Sample Count"
          }
        },
        "type": "object",
        "required": [
          "index",
          "start_ts",
          "end_ts",
          "start_lat",
          "start_lon",
          "end_lat",
          "end_lon",
          "distance_km",
          "duration_min",
          "moving_time_min",
          "idle_time_min",
          "max_speed_kmh",
          "sample_count"
        ],
        "title": "TripOut",
        "description": "One detected trip with derived mobility stats.\n\nTimestamps are ``None`` when the underlying samples carried none;\neverything else is zero-safe so the UI can render without guarding."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}